Fix errors and warnings

+++ TextureStreamer/src/Drawing/Drawers/RandomPointPicker.cpp
@@ -10,10 +10,10 @@ std::list<Point>::const_iterator RandomPointPicker::PickPoint(const std::list<Po
 {
 	if(points.size() == 0) return points.cend();
 
-	std::uniform_int_distribution<unsigned int> pointsLengthDistribution(0, points.size()-1);
+	std::uniform_int_distribution<size_t> pointsLengthDistribution(0, points.size()-1);
 
 	std::list<Point>::const_iterator it = points.cbegin();
-	for(unsigned int i = 0; i < pointsLengthDistribution(*m_generator); ++i)
+	for(size_t i = 0; i < pointsLengthDistribution(*m_generator); ++i)
 		it++;
 
 	return it;
